Add extract_playlist_insights
